agriguard-ml-pipeline/requirements.txt
absl-py==2.3.1
ai-edge-litert==1.3.0
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
astunparse==1.6.3
backports.strenum==1.2.8
certifi==2025.6.15
charset-normalizer==3.4.2
chex==0.1.89
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
etils==1.12.2
eval_type_backport==0.2.2
filelock==3.13.1
flatbuffers==25.2.10
flax==0.10.7
fonttools==4.58.5
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
humanfriendly==10.0
humanize==4.12.3
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.4
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.1
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
onnx==1.17.0
onnx-tf==1.6.0
onnx2tf==1.28.1
onnx_graphsurgeon==0.5.8
onnxruntime==1.22.0
onnxslim==0.1.59
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optax==0.2.5
optree==0.16.0
orbax-checkpoint==0.11.18
packaging==23.2
pandas==2.3.0
pillow==11.0.0
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
setuptools==70.2.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sng4onnx==1.0.4
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
tensorflow_decision_forests==1.12.0
tensorflowjs==4.22.0
tensorstore==0.1.76
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tifffile==2025.6.11
toolz==1.0.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchcam==0.4.0
torchvision==0.22.1+cu118
tqdm==4.67.1
treescope==0.1.9
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.162
ultralytics-thop==2.0.14
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wurlitzer==3.1.1
ydf==0.12.0
zipp==3.23.0
